No issue: move TestStrictModeManager to perf sub package.
