+++ app/src/test/java/org/mozilla/fenix/helpers/perf/TestStrictModeManager.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.helpers
+package org.mozilla.fenix.helpers.perf
 
 import android.os.StrictMode
 import io.mockk.mockk
